+++ test/view_component/view_component_test.rb
@@ -523,4 +523,14 @@ class ViewComponentTest < ViewComponent::TestCase
 
     assert_match(/MissingDefaultCollectionParameterComponent initializer must accept `missing_default_collection_parameter` collection parameter/, exception.message)
   end
+
+  def test_collection_component_with_trailing_comma_attr_reader
+    exception = assert_raises ArgumentError do
+      render_inline(
+        ProductReaderOopsComponent.with_collection(["foo"])
+      )
+    end
+
+    assert_match(/ProductReaderOopsComponent initializer is empty or invalid/, exception.message)
+  end
 end